+++ pom.xml
@@ -24,20 +24,9 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <jetty.version>9.2.9.v20150224</jetty.version>
-    <atmosphere.version>2.2.0</atmosphere.version>
   </properties>
 
   <dependencies>
-    <dependency>
-      <groupId>org.atmosphere</groupId>
-      <artifactId>atmosphere-runtime</artifactId>
-      <version>${atmosphere.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.atmosphere</groupId>
-      <artifactId>atmosphere-annotations</artifactId>
-      <version>${atmosphere.version}</version>
-    </dependency>
     <dependency>
       <groupId>org.eclipse.jetty.websocket</groupId>
       <artifactId>websocket-server</artifactId>
