+++ src/main/java/com/cylentsystems/dropwizard/sisu/Websocket.java
@@ -1,4 +1,4 @@
-package com.cylentsystems.dropwizard.sisu.bundles;
+package com.cylentsystems.dropwizard.sisu;
 
 /**
  * *****************************************************************************
